+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 
 # Unreleased
 
-- [`Breaking`] Removes private key from the Account class to support keyless and passkey
+- [`Breaking`] Removes private key from the Account class to support MultiKey accounts.
+- [`Breaking`] Removes the `sign` function from transactionBuilder.ts.  Use `Account.signTransactionWithAuthenticator` instead.
 - Refactors the core/accounts folder to the top level
 - Separates the signing message functionality out of the transactionSubmission.ts file
 - Adds an Account implementation for MultiKey accounts